backend/requirements.txt
fastapi
uvicorn
pydantic
langchain-ollama
sqlalchemy
google-cloud-texttospeech
google-cloud-speech
python-multipart
